fix(map_methods): change type of mask to int32

+++ makiflow/generators/segmentator/gen_layers.py
@@ -118,7 +118,7 @@ class InputGenNumpyGetterLayer(GenLayer):
             gen,
             output_types={
                 SegmentPathGenerator.IMAGE: tf.float32,
-                SegmentPathGenerator.MASK: tf.float32
+                SegmentPathGenerator.MASK: tf.int32
             }
         )
